# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText
  description: |
    This is the API documentation for WASAText.
  version: 2.1.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
...
paths:
...
  /conversations/{conversation_id}/messages:
    post:
      tags: 
        - Messages
        - Conversations
      operationId: sendMessage
      summary: Send a message in a conversation
      description: |
        Sends a new message to the specified conversation. If no conversation exists, a new one will be created.
      parameters:
        - name: conversation_id
          in: path
          required: true
          description: ID of the conversation
          schema: { $ref: "#/components/schemas/Id"}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewMessage"     
      responses:
        '201':
          description: Message successfully sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
              example:
                id: 987
                timestamp: 2025-08-03T16:55:00Z
                sender: "bob"
                type: "text"
                text: "Fine, thank you!"
                comments_count: 0
        '400': { $ref: "#/components/responses/Unauthorized" }
  /conversations/{conversation_id}/messages/{message_id}:
    parameters:
      - name: conversation_id
        in: path
        required: true
        description: ID of the conversation
        schema: 
          $ref: "#/components/schemas/Id"
      - name: message_id
        in: path
        required: true
        description: ID of the message to delete
        schema:
          $ref: "#/components/schemas/Id"
    post:
      tags:
        - Messages
        - Conversations
      operationId: forwardMessage
      summary: Forwards a message
      description: |
        Forwards a previously sent message to another conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload for the recipient conversation_id
              required:
                - recipient_conversation_id
              properties:
                recipient_conversation_id:
                  $ref: "#/components/schemas/Id"
                author:
                  $ref: "#/components/schemas/Username"
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: 987
                timestamp: 2025-08-03T16:55:00Z
                sender: "bob"
                type: "text"
                text: "Fine, thank you!"
                comments_count: 0
        '404': { $ref: "#/components/responses/NotFound" }
    delete:
      tags:
        - Messages
        - Conversations
      operationId: deleteMessage
      summary: Deletes a message
      description: |
        Deletes a previously sent message in a conversation.
      responses:
        '204':
          description: Message deleted successfully
        '404': { $ref: "#/components/responses/NotFound" }
  /conversations/{conversation_id}/messages/{message_id}/comments:
    parameters:
      - name: conversation_id
        in: path
        required: true
        description: ID of the conversation
        schema: 
          $ref: "#/components/schemas/Id"
      - name: message_id
        in: path
        required: true
        description: ID of the message to delete
        schema:
          $ref: "#/components/schemas/Id"
    post:
      tags:
        - Messages
        - Comments
      operationId: commentMessage
      summary: Adds a comment to a message
      description: Adds a new comment to the specified message in the given conversation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewComment"
      responses:
        '201':
          description: Comment added successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
              example:
                id: 42
                author: "alice123"
                text: "I totally agree with this!"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Messages
        - Comments
      operationId: uncommentMessage
      summary: Deletes a comment from a message
      description: Removes the specified comment from the message.
      responses:
        '204':
          description: Message deleted successfully
        '401': { $ref: "#/components/responses/Unauthorized" }
        '404': { $ref: "#/components/responses/NotFound" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Id:
      description:  Unique identifier for users, conversations, groups and messages
      type: integer
      format: int64
      minimum: 1
      example: 17
    Username:
      description: Name of a user
      type: string
      pattern: '^[a-zA-Z0-9_.-]{3,13}$'
      minLength: 3
      maxLength: 25
      example: "alice123"
...
    Pic:
      description: Base64-encoded image data, it can be JPEG, PNG, GIF, WebP
      type: string
      format: byte
      pattern: '^[A-Za-z0-9+/]+={0,2}$'
      minLength: 100
      maxLength: 500000
      example: "iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4//8/w38GIAXDIBKE0DHxgljNBAAO9TXL0Y4OHwAAAABJRU5ErkJggg=="
...
    Message:
      description: Message schema
      type: object
      required:
        - id
        - timestamp
        - sender
        - type
        - comments_count
      properties:
        id:
          $ref: "#/components/schemas/Id"
        timestamp:
          $ref: "#/components/schemas/Timestamp"
        sender:
          $ref: "#/components/schemas/Username" 
        type:
          type: string
          enum: [text, photo]
          description: Type of the message.
        comments_count:
          type: integer
          description: Number of comments (reactions) on the message.
          example: 2
        comments_authors:
          type: array
          description: List of users who have reacted to the message.
          minItems: 0
          maxItems: 1000
          items:
            $ref: "#/components/schemas/Username"
      oneOf:
        - required: [text]
          properties:
            text:
              type: string
              description: Text content of the message.
              pattern: '^.*$'
              minLength: 1
              maxLength: 1000
        - required: [photo]
          properties:
            photo:
              $ref: "#/components/schemas/Pic"
...
    NewMessage:
      type: object
      description: Message content to be sent by the client.
      required:
        - type
      properties:
        type:
          type: string
          description: Type of the message.
          enum: [text, photo]
          example: text
      oneOf:
        - required: [text]
          properties:
            text:
              type: string
              description: Text content of the message.
              pattern: '^.*$'
              minLength: 1
              maxLength: 1000
        - required: [photo]
          properties:
            photo:
              $ref: "#/components/schemas/Pic"
    Timestamp:
      type: string
      description: Date and time of message.
      format: date-time
      minLength: 20
      maxLength: 30
...
    Comment:
      type: object
      description: A comment (reaction) attached to a message.
      required:
        - id
        - author
        - text
      properties:
        id:
          $ref: "#/components/schemas/Id"
        author:
          $ref: "#/components/schemas/Username"
        text:
          type: string
          description: The content of the comment.
          minLength: 1
          maxLength: 300
    NewComment:
      type: object
      description: The comment (reaction) attached to a message.
      required: [text]
      properties:
        text:
          type: string
          description: The content of the comment.
          pattern: '^.*$'
          minLength: 1
          maxLength: 300
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g., missing or invalid fields).
    Unauthorized:
      description: The request requires user authentication or the provided credentials are invalid.
    NotFound:
      description: The requested resource was not found.
...